+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - Add rate limiting and a default ntfy.sh token support for ntfy notifications
+- Sort checks in daily/weekly/monthly reports by downtime duration, then by name (#1095)
 
 ## v3.11.2 - 2025-09-02